requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0.post1
openai>=1.40.0
pandas>=2.2.0